pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-pipelines"
version = "0.1.0"
description = "End-to-end document ingestion pipeline using LlamaIndex and OpenAI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyYAML==6.0.2",
    "jmespath==1.0.1",
    "llama-index==0.12.46",
    "llama-index-storage-index-store-postgres==0.4.0",
    "psycopg2==2.9.10",
    "llama-index-vector-stores-opensearch==0.5.6",
    "langchain-text-splitters==0.3.8",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pydantic-settings",
]

[tool.setuptools]
packages = ["ingestion"]
